cinereacapture/cincaptureweb/requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==24.4.2
Automat==24.8.1
Babel==2.10.3
bcc==0.29.1
bcrypt==3.2.2
blinker==1.7.0
Brlapi==0.8.5
certifi==2023.11.17
channels==4.2.0
chardet==5.2.0
click==8.1.6
cloud-init==24.4
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
constantly==23.10.4
cryptography==41.0.7
cupshelpers==1.0
daphne==4.1.2
dbus-python==1.3.2
defer==1.0.6
distro==1.9.0
distro-info==1.7+build1
Django==5.1.4
django-channels==0.7.0
django-crispy-forms==2.3
django-mathfilters==1.0.0
djangorestframework==3.15.2
duplicity==2.1.4
et_xmlfile==2.0.0
fasteners==0.18
httplib2==0.20.4
hyperlink==21.0.0
idna==3.6
incremental==24.7.2
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
louis==3.29.0
Mako==1.3.2.dev0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
monotonic==1.6
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.2
olefile==0.46
openpyxl==3.1.5
paramiko==2.12.0
pexpect==4.9.0
pillow==10.2.0
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.25.1
pycups==2.0.1
Pygments==2.17.2
PyGObject==3.48.2
PyJWT==2.7.0
PyNaCl==1.5.0
pyOpenSSL==24.3.0
pyparsing==3.1.1
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
service-identity==24.2.0
setuptools==75.6.0
six==1.16.0
sqlparse==0.5.3
systemd-python==235
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.10.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
urllib3==2.0.7
usb-creator==0.3.16
wadllib==1.3.6
xdg==5
xkit==0.0.0
zope.interface==7.2
